requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-phonenumber-field==7.0.0
phonenumberslite==8.12.55
sqlparse==0.4.2
